backend/requirements.txt
# ── OCR Engine ──
paddlepaddle
paddleocr
opencv-python
Pillow
numpy

# ── API Server ──
fastapi
uvicorn
python-multipart

# ── Vector Search (RAG) ──
sentence-transformers
faiss-cpu

# ── ML Classifier ──
scikit-learn
joblib

# ── Auth (OAuth2 + JWT) ──
PyJWT
httpx

# ── Testing ──
pytest
